+++ sandpack-client/src/client.ts
@@ -66,6 +66,8 @@ export interface ClientOptions {
     isFile: (path: string) => Promise<boolean>;
     readFile: (path: string) => Promise<string>;
   };
+
+  reactDevTools?: boolean;
 }
 
 export interface SandboxInfo {
@@ -273,6 +275,7 @@ export class SandpackClient {
       version: 3,
       isInitializationCompile,
       modules,
+      reactDevTools: this.options.reactDevTools,
       externalResources: this.options.externalResources || [],
       hasFileResolver: Boolean(this.options.fileResolver),
       disableDependencyPreprocessing: