+++ sandpack-react/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from "./useActiveCode";
-export * from "./useCodeSandboxLink";
 export * from "./useErrorMessage";
 export * from "./useLoadingOverlayState";
 export * from "./useSandpack";